+++ routes/register/diagramming/router_graduationsDiagramming.js
@@ -1,21 +1,28 @@
 module.exports = app => {
 
     const GraduationsDiagramming = require('../../../model').graduacoes_diagramadores;
+    const jwt = require('jsonwebtoken');
     let router = require('express').Router();
     const passport = require('passport');
     require('../../../config/passport')(passport);
     let getToken = require('../../../config/getToken');
+    let mySecret = 'TFMgQ29uc3VsdG9yaWEgJiBTaXN0ZW1hcyBMVERBIERFU0RFIDIwMTc='
     let token;
-    let tipoErro;
+    let typeError;
+    let errorAuth;
+    let successAuth;
     
     //Criando rotas de criação
     router.post('/',
     passport.authenticate('jwt', {
         session: false
     }), async function (req, res) {
-        token = await getToken(req.headers);
         try {
-            if (token) {
+            token = await getToken(req.headers);
+            jwt.verify(token, mySecret, function (err, data) {
+                (data ? (successAuth = Boolean(true)) : ((errorAuth = err) || (successAuth = Boolean(false))))
+            });
+            if (successAuth) { 
                 const graduacao_diagramadores_post = await  GraduationsDiagramming.create({
                     id: req.body.id,
                     diagramador_id: req.body.diagramador_id,
@@ -34,12 +41,12 @@ module.exports = app => {
         } catch (error) {
             console.error(error);
             if (error.name  ===  "SequelizeUniqueConstraintError") {
-                tipoErro = 'Já existe cadastro, ' + error.parent.constraint + ' não pode duplicar cadastro.'               
+                typeError = 'Já existe cadastro, ' + error.parent.constraint + ' não pode duplicar cadastro.'               
             }
             res.status(404).send(JSON.stringify({
                     "full_erro": error,
                     "error_detalhado": error.parent,
-                    "tipo_error": tipoErro,
+                    "tipo_error": typeError,
                     "status": false
                 }));            
         };
@@ -50,9 +57,12 @@ module.exports = app => {
     passport.authenticate('jwt', {
         session: false
     }), async function (req, res) {
-        token = await getToken(req.headers);
         try {
-            if (token) {
+            token = await getToken(req.headers);
+            jwt.verify(token, mySecret, function (err, data) {
+                (data ? (successAuth = Boolean(true)) : ((errorAuth = err) || (successAuth = Boolean(false))))
+            });
+            if (successAuth) { 
                 const graduacao_diagramadores_put = await  GraduationsDiagramming.update({
                     diagramador_id: req.body.diagramador_id,
                     graduacoes_id: req.body.graduacoes_id,
@@ -82,16 +92,27 @@ module.exports = app => {
 
     //Rota para Delete de Dados
     router.delete('/:diagramador_id', async function (req, res) {
-        try {            
-            const graduacao_diagramadores_delete = await GraduationsDiagramming.destroy({
-                where: {
-                        diagramador_id: req.params.diagramador_id
-                    }
-            });              
-            res.status(200).send(JSON.stringify({
-                "full_data": graduacao_diagramadores_delete,
-                "status": Boolean(true),
-            }));
+        try {
+            token = await getToken(req.headers);
+            jwt.verify(token, mySecret, function (err, data) {
+                (data ? (successAuth = Boolean(true)) : ((errorAuth = err) || (successAuth = Boolean(false))))
+            });
+            if (successAuth) { 
+                const graduacao_diagramadores_delete = await GraduationsDiagramming.destroy({
+                    where: {
+                            diagramador_id: req.params.diagramador_id
+                        }
+                });              
+                res.status(200).send(JSON.stringify({
+                    "full_data": graduacao_diagramadores_delete,
+                    "status": Boolean(true),
+                }));
+            } else {
+                res.status(401).send(JSON.stringify({
+                    "messagem": 'Senha não foi reconhecida.',
+                    "status": Boolean(false),
+                }));                
+            };
         } catch (error) {
             console.error(error);
             console.log(req.statusCode);
@@ -105,13 +126,24 @@ module.exports = app => {
 
     //Rota para Busca de Dados
     router.get('/', async function (req, res) {
-        try {            
-            const graduacao_diagramadores_get = await GraduationsDiagramming.findAll({
-                order: [
-                    ['id', 'ASC']
-                ]
-            });   
-            res.status(200).send(graduacao_diagramadores_get);
+        try {
+            token = await getToken(req.headers);
+            jwt.verify(token, mySecret, function (err, data) {
+                (data ? (successAuth = Boolean(true)) : ((errorAuth = err) || (successAuth = Boolean(false))))
+            });
+            if (successAuth) { 
+                const graduacao_diagramadores_get = await GraduationsDiagramming.findAll({
+                    order: [
+                        ['id', 'ASC']
+                    ]
+                });   
+                res.status(200).send(graduacao_diagramadores_get);
+            } else {
+                res.status(401).send(JSON.stringify({
+                    "messagem": 'Senha não foi reconhecida.',
+                    "status": Boolean(false),
+                }));                
+            };
         } catch (error) {
             console.error(error);
             console.log(req.statusCode);
@@ -125,13 +157,24 @@ module.exports = app => {
 
     //Rota para Busca de Dados
     router.get('/:id', async function (req, res) {
-        try {            
-            const graduacao_diagramadores_get = await GraduationsDiagramming.findAll({
-                where: {
-                        id: req.params.id
-                    }
-            });   
-            res.status(200).send(graduacao_diagramadores_get);
+        try {
+            token = await getToken(req.headers);
+            jwt.verify(token, mySecret, function (err, data) {
+                (data ? (successAuth = Boolean(true)) : ((errorAuth = err) || (successAuth = Boolean(false))))
+            });
+            if (successAuth) {         
+                const graduacao_diagramadores_get = await GraduationsDiagramming.findAll({
+                    where: {
+                            id: req.params.id
+                        }
+                });   
+                res.status(200).send(graduacao_diagramadores_get);
+            } else {
+                res.status(401).send(JSON.stringify({
+                    "messagem": 'Senha não foi reconhecida.',
+                    "status": Boolean(false),
+                }));                
+            };
         } catch (error) {
             console.error(error);
             console.log(req.statusCode);
